tool2/requirements.txt
# Core graph processing
networkx==2.8
numpy>=1.21.0

# Machine learning and similarity
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
transformers>=4.0.0
openai>=1.0.0

# File handling and threading
filelock>=3.0.0

# HTTP requests for LLM API
requests>=2.25.0

# Optional: For better performance
faiss-cpu>=1.7.0  # Optional: For faster similarity search
torch>=1.9.0      # Required by sentence-transformers

# Development and testing
pytest>=6.0.0     # For testing
pytest-cov>=3.0.0 # For coverage
black>=22.0.0     # For code formatting
flake8>=4.0.0     # For linting

python-dotenv>=0.19.0